requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
matplotlib>=3.7,<4.0
PyYAML>=6.0,<7.0
pytest>=8.0,<9.0
